# server/openapi.yaml
openapi: 3.0.3
info:
  title: ActivityOS API
  version: 0.31.0
  description: |
    Express + Firestore backend for ActivityOS — a **multi-tenant** platform
    for children's activity providers.

    ## v0.31.0 — Cross-provider marketplace (opt-in)

    The parent Browse feed becomes a real marketplace instead of being scoped
    (client-side) to providers a family had already booked.
    - **`GET /api/listings`** (no `?tenantId=`, no `?mine=1`) now returns the
      live/public/unarchived listings of every provider **that has opted in**
      (`settings.marketplaceListed === true`) — **plus**, for a signed-in
      parent, their own providers (anyone they've booked) regardless of opt-in.
    - A provider's **own storefront** (`GET /api/listings?tenantId=`) is
      unchanged — their public listings always show there whether or not
      they're in the marketplace.
    - New tenant setting **`marketplaceListed`** (in the library `settings`
      bag), default **false** — a provider opts in from Setup → Bookings →
      Marketplace. Opt-in is a deliberate choice; nobody is publicly listed
      without it.
    (Frontend: BrowseApp drops the old single-provider scoping and shows the
    server feed as-is.)

    ## v0.30.0 — Account, Data & privacy, parent Accidents

    Clearing the shared/parent mock pages.
    - **`/api/account`** (every role) — the signed-in user's profile. `GET`
      returns `{email, name, phone, marketingConsent, role, tenantId}` (identity
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - firebaseIdToken: []
paths:
...
  /api/meals:
    get:
      tags: [meals]
      summary: The day's menu + dietary board
      parameters:
        - { name: date, in: query, required: false, schema: { type: string } }
        - { name: tenantId, in: query, required: false, description: Platform only, schema: { type: string } }
      responses:
        "200": { description: "{date, menu, allergens, sessions[] (children with allergies/dietary + alerts), summary}" }
        "403": { $ref: "#/components/responses/Forbidden" }
    # (see /api/meals/{date} for setting the menu)
  /api/meals/{date}:
    put:
      tags: [meals]
      summary: Set the day's menu (operators)
      parameters: [{ name: date, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [meals]
              properties:
                meals:
                  type: array
                  items:
                    type: object
                    properties:
                      id: { type: string }
                      type: { type: string }
                      description: { type: string }
                      allergens: { type: array, items: { type: string } }
      responses:
        "200": { description: Saved }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/invites:
    get:
      tags: [invites]
      summary: List my tenant's invites
      responses:
        "200":
          description: Invites, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invite" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [invites]
      summary: Create an invite (link, optionally emailed)
      description: |
        Who may invite whom: **company** → franchise or staff;
        **franchise / freelancer** → staff only. With `email` the invite is
        emailed straight to the recipient (and recorded as `sentTo`);
        without it, share the returned URL yourself. Either way the
        recipient signs up through it and joins the tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [franchise, staff] }
                email: { type: string, description: Send the invite to this address as well as returning the link }
      responses:
        "201":
          description: The invite
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  url: { type: string, example: /signup?invite=abc123 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/invites/{token}:
    get:
      tags: [invites]
      summary: Preview an invite (no auth — the token is the secret)
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: What the invite is for
          content:
            application/json:
              schema:
                type: object
                properties:
                  role: { type: string, enum: [franchise, staff] }
                  tenantName: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }
        "410":
          description: Invite already used
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/invites/{token}/accept:
    post:
      tags: [invites]
      summary: Join the tenant (after creating an account)
      description: |
        Sets the account's role + tenant from the invite and burns the
        token. 409 if the account already has a type.
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                type: object
                properties:
                  role: { type: string }
                  tenantId: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Account type already set
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "410":
          description: Invite already used
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token (see the Authentication section above).
...
  responses:
    BadRequest:
      description: Validation failed (`error` is a message or zod issue list)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing/invalid Firebase ID token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The account's role/tenant doesn't allow this
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found (or outside the caller's tenant scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          description: Human-readable message, or an array of zod issues on 400.
          oneOf:
            - type: string
            - type: array
              items: { type: object }
...
    Invite:
      type: object
      properties:
        token: { type: string }
        role: { type: string, enum: [franchise, staff] }
        createdAt: { type: string }
        usedBy: { type: string, nullable: true }